wc re-write
