+++ canflood/_pars/WebConnections.ini
@@ -2,4 +2,9 @@
 serverType=WCS
 url=http://preview.grid.unep.ch/geoserver/wcs?bbox=-180,-89,180,84&styles=&version=1.0.0&coverage=GAR2015:flood_hazard_1000_yrp&width=640&height=309&crs=EPSG:4326
 
+[ECCC_NationalPollutantReleaseInventory_NPRI]
+serverType=arcgisfeatureserver
+url=https://maps-cartes.ec.gc.ca/arcgis/rest/services/NPRI_INRP/NPRI_INRP/MapServer
+
+
