+++ canflood/_pars/WebConnections1.ini
@@ -1,0 +1,4 @@
+[GAR15]
+group=connections-wcs\GAR
+invertAxisOrientation=False
+url=http://preview.grid.unep.ch/geoserver/wcs?bbox=-180,-89,180,84&styles=&version=1.0.0&coverage=GAR2015:flood_hazard_1000_yrp&width=640&height=309&crs=EPSG:4326
